[gl] extracted some common GPU rendering code (polygon rendering functions)
